+++ app/src/test/java/com/pocketshell/app/composer/PromptComposerViewModelTest.kt
@@ -215,6 +215,9 @@ class PromptComposerViewModelTest {
         speechRecognitionProvider: PromptComposerViewModel.SpeechRecognitionProvider =
             UnavailableSpeechRecognitionProvider,
         composerDraftStore: ComposerDraftStore = DisabledComposerDraftStore,
+        pendingTranscriptionStore: PromptComposerViewModel.PendingTranscriptionQueue =
+            DisabledPendingTranscriptionQueue,
+        connectivity: PromptComposerViewModel.ConnectivityProbe = AlwaysOnlineConnectivityProbe,
         outboundQueueStore: OutboundQueueStore = DisabledOutboundQueueStore,
         outboundAttachmentSidecarStore: OutboundAttachmentSidecarStore? = null,
         savedStateHandle: SavedStateHandle = SavedStateHandle(),
@@ -226,6 +229,8 @@ class PromptComposerViewModelTest {
             apiKeyStorage = storage,
             voiceSettings = voiceSettings,
             speechRecognitionProvider = speechRecognitionProvider,
+            pendingTranscriptionStore = pendingTranscriptionStore,
+            connectivity = connectivity,
             composerDraftStore = composerDraftStore,
             outboundQueueStore = outboundQueueStore,
             outboundAttachmentSidecarStore = outboundAttachmentSidecarStore,
@@ -1753,728 +1758,6 @@ class PromptComposerViewModelTest {
     private class FakeConnectivity(initial: Boolean = true) : PromptComposerViewModel.ConnectivityProbe {
         var online: Boolean = initial
         override fun refresh(): Boolean = online
-    }
-
-    /**
-     * Issue #688: a [WhisperClient] that holds every transcribe call open
-     * until [releaseAll] is invoked. Lets the test make two retry triggers
-     * provably concurrent (both in-flight at the same time) before either
-     * resolves — the only way to reproduce the duplicate-insert race
-     * deterministically. [callCount] records how many round-trips actually
-     * fired, so the test can assert the in-flight guard deduped them.
-     */
-    private class GatedWhisperClient(
-        private val result: Result<String>,
-    ) : WhisperClient {
-        private val gate = CompletableDeferred<Unit>()
-
-        @Volatile
-        var callCount: Int = 0
-            private set
-
-        override suspend fun transcribe(audio: ByteArray, language: String?): Result<String> {
-            callCount++
-            gate.await()
-            return result
-        }
-
-        fun releaseAll() {
-            if (!gate.isCompleted) gate.complete(Unit)
-        }
-    }
-
-    private fun newVmWithQueue(
-        mic: PromptComposerViewModel.MicCapture = FakeMicCapture(),
-        whisper: WhisperClient? = fakeWhisperClient { Result.success("hello world") },
-        storage: ApiKeyVault = newStorage().also { it.save("sk-test".toCharArray()) },
-        samplerDispatcher: TestDispatcher? = null,
-        clock: () -> Long = { System.currentTimeMillis() },
-        voiceSettings: PromptComposerViewModel.VoiceSettingsSnapshot = FakeVoiceSettings(),
-        queue: FakePendingQueue = FakePendingQueue(),
-        connectivity: FakeConnectivity = FakeConnectivity(initial = true),
-    ): Pair<PromptComposerViewModel, FakePendingQueue> {
-        val factory = WhisperClientFactory { whisper }
-        val vm = PromptComposerViewModel(
-            audioRecorder = mic,
-            whisperClientFactory = factory,
-            apiKeyStorage = storage,
-            voiceSettings = voiceSettings,
-            pendingTranscriptionStore = queue,
-            connectivity = connectivity,
-        )
-        if (samplerDispatcher != null) vm.samplerDispatcher = samplerDispatcher
-        vm.clock = clock
-        // Issue #882: track for teardown so a test that leaves recording active
-        // doesn't leak the #880 ticker loop into `runTest`'s final
-        // `advanceUntilIdle` (which would hang the suite). See
-        // [tearDownViewModels]. Mirrors [newVm].
-        createdViewModels += vm
-        // Issue #891: disable the overall-send watchdog by default (see [newVm]).
-        vm.setSendWatchdogTimeoutForTest(null)
-        return vm to queue
-    }
-
-    @Test
-    fun audioIsPersistedBeforeWhisperCall() = runTest {
-        val (vm, queue) = newVmWithQueue(
-            whisper = fakeWhisperClient { Result.success("hello") },
-            samplerDispatcher = StandardTestDispatcher(testScheduler),
-        )
-        vm.onMicTap()
-        runCurrent()
-        vm.onMicTap()
-        advanceUntilIdle()
-
-        // Persisted exactly once before the Whisper round-trip.
-        assertEquals(1, queue.enqueueCount)
-        // Whisper success deleted the row.
-        assertEquals(listOf("pending-1"), queue.succeededIds)
-        // Draft has the transcript.
-        assertEquals("hello", vm.uiState.value.draft)
-    }
-
-    @Test
-    fun whisperFailureKeepsAudioOnDiskAndBumpsRetryCount() = runTest {
-        val (vm, queue) = newVmWithQueue(
-            whisper = fakeWhisperClient {
-                Result.failure(WhisperException.Transport("offline"))
-            },
-            samplerDispatcher = StandardTestDispatcher(testScheduler),
-        )
-        vm.onMicTap()
-        runCurrent()
-        vm.onMicTap()
-        advanceUntilIdle()
-
-        assertEquals(1, queue.enqueueCount)
-        assertEquals(0, queue.succeededIds.size)
-        assertEquals(1, queue.failureIds.size)
-        assertEquals("pending-1", queue.failureIds[0].first)
-        // Error banner mirrors the failure reason.
-        assertNotNull(vm.uiState.value.error)
-    }
-
-    @Test
-    fun networkTimeoutLeavesTranscribingBeforePendingFailureWriteCompletes() = runTest {
-        val queue = FakePendingQueue()
-        queue.markFailureStarted = CompletableDeferred()
-        queue.markFailureRelease = CompletableDeferred()
-        val whisperRelease = CompletableDeferred<Unit>()
-        val attachmentPath = "~/.pocketshell/attachments/host-1/keep-this.png"
-        val whisper = object : WhisperClient {
-            override suspend fun transcribe(audio: ByteArray, language: String?): Result<String> {
-                whisperRelease.await()
-                return Result.failure(WhisperException.Transport("timeout"))
-            }
-        }
-        val (vm, _) = newVmWithQueue(
-            whisper = whisper,
-            samplerDispatcher = StandardTestDispatcher(testScheduler),
-            queue = queue,
-        )
-
-        vm.onDraftChange("keep this draft")
-        vm.seedAttachment(attachmentPath)
-        vm.onMicTap()
-        runCurrent()
-        vm.onMicTap()
-        assertEquals(RecordingState.Transcribing, vm.uiState.value.recording)
-
-        whisperRelease.complete(Unit)
-        runCurrent()
-        queue.markFailureStarted!!.await()
-
-        assertEquals(RecordingState.Idle, vm.uiState.value.recording)
-        assertEquals("Network error: timeout", vm.uiState.value.error)
-        assertEquals("keep this draft", vm.uiState.value.draft)
-        assertEquals(
-            listOf(
-                PromptComposerViewModel.StagedAttachment(
-                    remotePath = attachmentPath,
-                    displayName = "keep-this.png",
-                ),
-            ),
-            vm.uiState.value.attachments,
-        )
-        assertNotNull(queue.storedAudio("pending-1"))
-        assertTrue(queue.failureIds.isEmpty())
-
-        queue.markFailureRelease!!.complete(Unit)
-        advanceUntilIdle()
-
-        assertEquals(
-            listOf("pending-1" to "Network error: timeout"),
-            queue.failureIds,
-        )
-        assertNotNull(queue.storedAudio("pending-1"))
-    }
-
-    @Test
-    fun offlineSkipsWhisperAndQueuesDirectly() = runTest {
-        var whisperCalls = 0
-        val whisper = object : WhisperClient {
-            override suspend fun transcribe(audio: ByteArray, language: String?): Result<String> {
-                whisperCalls++
-                return Result.success("never")
-            }
-        }
-        val (vm, queue) = newVmWithQueue(
-            whisper = whisper,
-            samplerDispatcher = StandardTestDispatcher(testScheduler),
-            connectivity = FakeConnectivity(initial = false),
-        )
-        vm.onMicTap()
-        runCurrent()
-        vm.onMicTap()
-        advanceUntilIdle()
-
-        assertEquals(1, queue.enqueueCount)
-        assertEquals(0, whisperCalls)
-        // Banner reads "waiting for network".
-        assertEquals(
-            com.pocketshell.app.voice.PendingTranscriptionItem.NETWORK_WAITING_MESSAGE,
-            vm.uiState.value.error,
-        )
-    }
-
-    @Test
-    fun marginalNonSilentGuardRejectionIsQueuedForRetryInsteadOfLost() = runTest {
-        // Regression for #587: a real, long capture can land below the
-        // conservative speech RMS floor. It must still skip Whisper on the
-        // first pass (#452), but the audio should be preserved for retry/export
-        // instead of being discarded under an unrecoverable no-speech banner.
-        val marginalAudio = SpeechAudioGuard.speechWavForTesting(durationMs = 1_200, amplitude = 0.004f)
-        assertFalse(SpeechAudioGuard.hasSpeechEnergy(marginalAudio))
-        assertTrue(SpeechAudioGuard.isRecoverableNoSpeechRejection(marginalAudio))
-
-        var whisperCalls = 0
-        val whisper = object : WhisperClient {
-            override suspend fun transcribe(audio: ByteArray, language: String?): Result<String> {
-                whisperCalls++
-                return Result.success("should not be called")
-            }
-        }
-        val (vm, queue) = newVmWithQueue(
-            mic = FakeMicCapture(audio = marginalAudio),
-            whisper = whisper,
-            samplerDispatcher = StandardTestDispatcher(testScheduler),
-        )
-
-        vm.onMicTap()
-        runCurrent()
-        vm.onMicTap()
-        advanceUntilIdle()
-
-        assertEquals(0, whisperCalls)
-        assertEquals(1, queue.enqueueCount)
-        assertTrue(queue.storedAudio("pending-1")!!.contentEquals(marginalAudio))
-        assertEquals(
-            PromptComposerViewModel.SUSPICIOUS_NO_SPEECH_RETRY_MESSAGE,
-            queue.snapshot().single().lastErrorMessage,
-        )
-        assertEquals(
-            PromptComposerViewModel.SUSPICIOUS_NO_SPEECH_RETRY_MESSAGE,
-            vm.uiState.value.error,
-        )
-        assertEquals(RecordingState.Idle, vm.uiState.value.recording)
-    }
-
-    @Test
-    fun pureSilenceStillSkipsWhisperAndIsNotQueued() = runTest {
-        var whisperCalls = 0
-        val whisper = object : WhisperClient {
-            override suspend fun transcribe(audio: ByteArray, language: String?): Result<String> {
-                whisperCalls++
-                return Result.success("시청해주셔서 감사합니다!")
-            }
-        }
-        val (vm, queue) = newVmWithQueue(
-            mic = SilentCaptureMic(),
-            whisper = whisper,
-            samplerDispatcher = StandardTestDispatcher(testScheduler),
-        )
-
-        vm.onMicTap()
-        runCurrent()
-        vm.onMicTap()
-        advanceUntilIdle()
-
-        assertEquals(0, whisperCalls)
-        assertEquals(0, queue.enqueueCount)
-        assertEquals(
-            PromptComposerViewModel.NO_SPEECH_DETECTED_MESSAGE,
-            vm.uiState.value.error,
-        )
-    }
-
-    @Test
-    fun emptyWhisperSuccessKeepsRealRecordingQueuedForRetry() = runTest {
-        val (vm, queue) = newVmWithQueue(
-            mic = FakeMicCapture(audio = SpeechAudioGuard.speechWavForTesting()),
-            whisper = fakeWhisperClient { Result.success("   \n") },
-            samplerDispatcher = StandardTestDispatcher(testScheduler),
-        )
-        vm.onDraftChange("typed prompt")
-
-        vm.onMicTap()
-        runCurrent()
-        vm.onMicTap()
-        advanceUntilIdle()
-
-        assertEquals(1, queue.enqueueCount)
-        assertEquals(emptyList<String>(), queue.succeededIds)
-        assertEquals(
-            listOf("pending-1" to PromptComposerViewModel.EMPTY_TRANSCRIPTION_RETRY_MESSAGE),
-            queue.failureIds,
-        )
-        assertNotNull(queue.storedAudio("pending-1"))
-        assertEquals("typed prompt", vm.uiState.value.draft)
-        assertEquals(
-            PromptComposerViewModel.EMPTY_TRANSCRIPTION_RETRY_MESSAGE,
-            vm.uiState.value.error,
-        )
-    }
-
-    @Test
-    fun hallucinationWhisperSuccessKeepsSpeechLikeRecordingQueuedForRetry() = runTest {
-        val (vm, queue) = newVmWithQueue(
-            mic = FakeMicCapture(audio = SpeechAudioGuard.speechWavForTesting()),
-            whisper = fakeWhisperClient { Result.success("시청해주셔서 감사합니다!") },
-            samplerDispatcher = StandardTestDispatcher(testScheduler),
-        )
-        vm.onDraftChange("typed prompt")
-
-        vm.onMicTap()
-        runCurrent()
-        vm.onMicTap()
-        advanceUntilIdle()
-
-        assertEquals(1, queue.enqueueCount)
-        assertEquals(emptyList<String>(), queue.succeededIds)
-        assertEquals(
-            listOf("pending-1" to PromptComposerViewModel.NO_SPEECH_DETECTED_MESSAGE),
-            queue.failureIds,
-        )
-        assertNotNull(queue.storedAudio("pending-1"))
-        assertEquals("typed prompt", vm.uiState.value.draft)
-        assertEquals(
-            PromptComposerViewModel.NO_SPEECH_DETECTED_MESSAGE,
-            vm.uiState.value.error,
-        )
-    }
-
-    @Test
-    fun retryPendingSuccessAppendsToDraftAndClearsQueue() = runTest {
-        // Seed a queued item without going through the recording path.
-        val queue = FakePendingQueue()
-        queue.nextId = "retryable"
-        queue.enqueueAudio(ByteArray(8) { 1 }, "composer")
-        // Bump it once to simulate a prior failure (so the retry path is
-        // hit, not the initial-attempt path).
-        queue.markFailure("retryable", "Whisper auth failed")
-
-        val (vm, _) = newVmWithQueue(
-            whisper = fakeWhisperClient { Result.success("retry transcript") },
-            samplerDispatcher = StandardTestDispatcher(testScheduler),
-            queue = queue,
-        )
-        vm.retryPending("retryable")
-        advanceUntilIdle()
-
-        assertEquals("retry transcript", vm.uiState.value.draft)
-        assertTrue(queue.succeededIds.contains("retryable"))
-    }
-
-    @Test
-    fun retryPendingFailureBumpsRetryCountAndSurfacesError() = runTest {
-        val queue = FakePendingQueue()
-        queue.nextId = "retryable"
-        queue.enqueueAudio(ByteArray(8) { 1 }, "composer")
-
-        val (vm, _) = newVmWithQueue(
-            whisper = fakeWhisperClient { Result.failure(WhisperException.Server("oops", 500)) },
-            samplerDispatcher = StandardTestDispatcher(testScheduler),
-            queue = queue,
-        )
-        vm.retryPending("retryable")
-        advanceUntilIdle()
-
-        assertEquals(1, queue.failureIds.size)
-        assertEquals("retryable", queue.failureIds[0].first)
-        assertNotNull(vm.uiState.value.error)
-    }
-
-    @Test
-    fun retryPendingEmptyWhisperResponseKeepsAudioQueued() = runTest {
-        val queue = FakePendingQueue()
-        queue.nextId = "retryable-empty"
-        queue.enqueueAudio(SpeechAudioGuard.speechWavForTesting(), "composer")
-
-        val (vm, _) = newVmWithQueue(
-            whisper = fakeWhisperClient { Result.success(" \n ") },
-            samplerDispatcher = StandardTestDispatcher(testScheduler),
-            queue = queue,
-        )
-        vm.retryPending("retryable-empty")
-        advanceUntilIdle()
-
-        assertEquals(
-            listOf("retryable-empty" to PromptComposerViewModel.EMPTY_TRANSCRIPTION_RETRY_MESSAGE),
-            queue.failureIds,
-        )
-        assertFalse(queue.succeededIds.contains("retryable-empty"))
-        assertNotNull(queue.storedAudio("retryable-empty"))
-        assertEquals("", vm.uiState.value.draft)
-        assertEquals(
-            PromptComposerViewModel.EMPTY_TRANSCRIPTION_RETRY_MESSAGE,
-            vm.uiState.value.error,
-        )
-    }
-
-    @Test
-    fun retryPendingAtCapShortCircuits() = runTest {
-        val queue = FakePendingQueue()
-        queue.nextId = "capped"
-        queue.enqueueAudio(ByteArray(8) { 1 }, "composer")
-        // Bump to the cap.
-        queue.markFailure("capped", "1")
-        queue.markFailure("capped", "2")
-        queue.markFailure("capped", "3")
-
-        val (vm, _) = newVmWithQueue(
-            whisper = fakeWhisperClient { Result.success("never") },
-            samplerDispatcher = StandardTestDispatcher(testScheduler),
-            queue = queue,
-        )
-        val priorFailures = queue.failureIds.size
-        val priorSuccesses = queue.succeededIds.size
-        vm.retryPending("capped")
-        advanceUntilIdle()
-        // Should not have called Whisper, so neither counter moved.
-        assertEquals(priorFailures, queue.failureIds.size)
-        assertEquals(priorSuccesses, queue.succeededIds.size)
-    }
-
-    @Test
-    fun retryPendingWhileOfflineSkipsWhisperAndNudgesUser() = runTest {
-        val queue = FakePendingQueue()
-        queue.nextId = "offline-retry"
-        queue.enqueueAudio(ByteArray(8) { 1 }, "composer")
-
-        var whisperCalls = 0
-        val whisper = object : WhisperClient {
-            override suspend fun transcribe(audio: ByteArray, language: String?): Result<String> {
-                whisperCalls++
-                return Result.success("never")
-            }
-        }
-        val (vm, _) = newVmWithQueue(
-            whisper = whisper,
-            samplerDispatcher = StandardTestDispatcher(testScheduler),
-            queue = queue,
-            connectivity = FakeConnectivity(initial = false),
-        )
-        vm.retryPending("offline-retry")
-        advanceUntilIdle()
-
-        assertEquals(0, whisperCalls)
-        assertEquals(
-            com.pocketshell.app.voice.PendingTranscriptionItem.NETWORK_WAITING_MESSAGE,
-            vm.uiState.value.error,
-        )
-    }
-
-    @Test
-    fun discardPendingClearsQueueEntry() = runTest {
-        val queue = FakePendingQueue()
-        queue.nextId = "to-discard"
-        queue.enqueueAudio(ByteArray(8) { 1 }, "composer")
-        val (vm, _) = newVmWithQueue(queue = queue)
-        vm.discardPending("to-discard")
-        advanceUntilIdle()
-        assertTrue(queue.discardedIds.contains("to-discard"))
-    }
-
-    // -- Issue #688: retry idempotency + clear-pending state ----------------
-
-    /**
-     * Issue #688 regression: a flaky-network retry must insert the
-     * transcript EXACTLY ONCE no matter how many retry triggers overlap.
-     *
-     * Reproduces the maintainer's dogfood report: a recording fails
-     * (Network error: timeout) and lands in the pending queue marked
-     * "waiting for network". An auto-retry (foreground-resume) and a manual
-     * Retry tap then race the SAME row. Before the fix both round-trips ran
-     * to completion and BOTH appended the transcript → the draft was
-     * duplicated. After the fix the second trigger is a no-op while the
-     * first is in flight, so the transcript is inserted once.
-     *
-     * The [GatedWhisperClient] holds every transcribe call open until the
-     * test releases it, so the two triggers are provably concurrent.
-     */
-    @Test
-    fun overlappingAutoRetryAndManualRetryInsertTranscriptExactlyOnce() = runTest {
-        val queue = FakePendingQueue()
-        queue.nextId = "flaky-1"
-        // Queued as "waiting for network" so onForegroundResume auto-retries it.
-        queue.enqueueAudio(
-            ByteArray(8) { 1 },
-            "composer",
-            initialError = com.pocketshell.app.voice.PendingTranscriptionItem.NETWORK_WAITING_MESSAGE,
-        )
-
-        val whisper = GatedWhisperClient(result = Result.success("hello world"))
-        val (vm, _) = newVmWithQueue(
-            whisper = whisper,
-            samplerDispatcher = StandardTestDispatcher(testScheduler),
-            queue = queue,
-            connectivity = FakeConnectivity(initial = true),
-        )
-
-        // Trigger 1: foreground-resume auto-retry for the waiting row.
-        vm.onForegroundResume()
-        runCurrent()
-        // The first round-trip is now open (gated). Trigger 2: the user taps
-        // Retry on the very same row while the auto-retry is still in flight.
-        vm.retryPending("flaky-1")
-        runCurrent()
-
-        // Let every gated transcribe call resolve.
-        whisper.releaseAll()
-        advanceUntilIdle()
-
-        // Exactly one insertion — the draft is NOT "hello world hello world".
-        assertEquals("hello world", vm.uiState.value.draft)
-        // The round-trip ran at most once for this row (the duplicate
-        // trigger was deduped, not just deduped at insert time).
-        assertEquals(1, whisper.callCount)
-        assertTrue(queue.succeededIds.contains("flaky-1"))
-        // The pending queue is fully cleared — no stale row remains.
-        assertTrue(queue.snapshot().isEmpty())
-    }
-
-    /**
-     * Issue #688: a late success for a row that another retry already
-     * inserted (or the user discarded) must be DROPPED, not appended a
-     * second time. Simulates the "timeout that actually succeeded
-     * server-side" the report calls out: the row is gone by the time the
-     * straggler round-trip returns, so its transcript must not land.
-     */
-    @Test
-    fun lateSuccessForAlreadyResolvedRowIsDropped() = runTest {
-        val queue = FakePendingQueue()
-        queue.nextId = "straggler"
-        queue.enqueueAudio(ByteArray(8) { 1 }, "composer")
-        queue.markFailure("straggler", "Network error: timeout")
-
-        val whisper = GatedWhisperClient(result = Result.success("late transcript"))
-        val (vm, _) = newVmWithQueue(
-            whisper = whisper,
-            samplerDispatcher = StandardTestDispatcher(testScheduler),
-            queue = queue,
-            connectivity = FakeConnectivity(initial = true),
-        )
-
-        vm.retryPending("straggler")
-        runCurrent()
-        // While the retry round-trip is open, the user discards the row
-        // (or a sibling resolved it). The store no longer has it.
-        queue.markSucceeded("straggler")
-        runCurrent()
-
-        // Now the straggler round-trip returns success.
-        whisper.releaseAll()
-        advanceUntilIdle()
-
-        // The transcript must NOT be inserted — the row was already gone.
-        assertEquals("", vm.uiState.value.draft)
-        assertTrue(queue.snapshot().isEmpty())
-    }
-
-    /**
-     * Issue #688: while a retry round-trip is in flight, the row must
-     * advertise a visible "retrying" state so the user gets immediate
-     * feedback rather than a dead button. Once it resolves the flag clears.
-     */
-    @Test
-    fun retryPublishesRetryingFeedbackWhileInFlight() = runTest {
-        val queue = FakePendingQueue()
-        queue.nextId = "feedback-1"
-        queue.enqueueAudio(ByteArray(8) { 1 }, "composer")
-        queue.markFailure("feedback-1", "Network error: timeout")
-
-        val whisper = GatedWhisperClient(result = Result.success("done"))
-        val (vm, _) = newVmWithQueue(
-            whisper = whisper,
-            samplerDispatcher = StandardTestDispatcher(testScheduler),
-            queue = queue,
-            connectivity = FakeConnectivity(initial = true),
-        )
-
-        vm.retryPending("feedback-1")
-        runCurrent()
-        // Mid-flight: the row is marked retrying.
-        assertTrue(vm.uiState.value.retryingIds.contains("feedback-1"))
-
-        whisper.releaseAll()
-        advanceUntilIdle()
-        // Resolved: the retrying flag is cleared.
-        assertFalse(vm.uiState.value.retryingIds.contains("feedback-1"))
-    }
-
-    /**
-     * Issue #688: a manual Retry tap while an auto-retry for the same row
-     * is already in flight must NOT start a second round-trip. The duplicate
-     * tap is a no-op (deduped by the in-flight guard).
-     */
-    @Test
-    fun manualRetryWhileAutoRetryInFlightDoesNotStack() = runTest {
-        val queue = FakePendingQueue()
-        queue.nextId = "no-stack"
-        queue.enqueueAudio(
-            ByteArray(8) { 1 },
-            "composer",
-            initialError = com.pocketshell.app.voice.PendingTranscriptionItem.NETWORK_WAITING_MESSAGE,
-        )
-
-        val whisper = GatedWhisperClient(result = Result.success("once"))
-        val (vm, _) = newVmWithQueue(
-            whisper = whisper,
-            samplerDispatcher = StandardTestDispatcher(testScheduler),
-            queue = queue,
-            connectivity = FakeConnectivity(initial = true),
-        )
-
-        vm.onForegroundResume()
-        runCurrent()
-        repeat(5) {
-            vm.retryPending("no-stack")
-            runCurrent()
-        }
-        whisper.releaseAll()
-        advanceUntilIdle()
-
-        assertEquals("only one transcribe call may fire for a row", 1, whisper.callCount)
-        assertEquals("once", vm.uiState.value.draft)
-    }
-
-    /**
-     * Issue #688: once the retry cap is hit, nothing auto-inserts. The
-     * foreground-resume sweep must not re-fire a capped row, and a manual
-     * Retry is gated at the cap.
-     */
-    @Test
-    fun cappedRowNeverAutoInsertsOnForegroundResume() = runTest {
-        val queue = FakePendingQueue()
-        queue.nextId = "capped-row"
-        // Cap via three failures while flagged as a network-waiting row so
-        // a naive auto-retry sweep would still try it.
-        queue.enqueueAudio(
-            ByteArray(8) { 1 },
-            "composer",
-            initialError = com.pocketshell.app.voice.PendingTranscriptionItem.NETWORK_WAITING_MESSAGE,
-        )
-        queue.markFailure("capped-row", "Network error: timeout")
-        queue.markFailure("capped-row", "Network error: timeout")
-        queue.markFailure("capped-row", "Network error: timeout")
-
-        var whisperCalls = 0
-        val whisper = object : WhisperClient {
-            override suspend fun transcribe(audio: ByteArray, language: String?): Result<String> {
-                whisperCalls++
-                return Result.success("should never insert")
-            }
-        }
-        val (vm, _) = newVmWithQueue(
-            whisper = whisper,
-            samplerDispatcher = StandardTestDispatcher(testScheduler),
-            queue = queue,
-            connectivity = FakeConnectivity(initial = true),
-        )
-
-        vm.onForegroundResume()
-        advanceUntilIdle()
-        vm.retryPending("capped-row")
-        advanceUntilIdle()
-
-        assertEquals("a capped row must not call Whisper", 0, whisperCalls)
-        assertEquals("", vm.uiState.value.draft)
-        // Row is still present in the explicit failed state, audio preserved.
-        assertTrue(queue.snapshot().any { it.id == "capped-row" && it.atRetryCap })
-        assertNotNull(queue.storedAudio("capped-row"))
-    }
-
-    @Test
-    fun savePendingAsAudioFileExportsAndSurfacesPath() = runTest {
-        val queue = FakePendingQueue()
-        queue.nextId = "to-save"
-        queue.enqueueAudio(ByteArray(8) { 1 }, "composer")
-        val (vm, _) = newVmWithQueue(queue = queue)
-        vm.savePendingAsAudioFile("to-save")
-        advanceUntilIdle()
-        assertTrue(queue.savedIds.contains("to-save"))
-        assertEquals(
-            "/data/files/voice-exports/to-save.wav",
-            vm.uiState.value.savedAudioPath,
-        )
-        vm.clearSavedAudioConfirmation()
-        assertNull(vm.uiState.value.savedAudioPath)
-    }
-
-    @Test
-    fun foregroundResumeAutoRetriesOnlyWaitingForNetworkItems() = runTest {
-        val queue = FakePendingQueue()
-        // Two queued rows: one waiting for network (auto-retry candidate),
-        // one already failed (manual-only).
-        queue.nextId = "offline-one"
-        queue.enqueueAudio(
-            ByteArray(8),
-            "composer",
-            initialError = com.pocketshell.app.voice.PendingTranscriptionItem.NETWORK_WAITING_MESSAGE,
-        )
-        queue.nextId = "failed-one"
-        queue.enqueueAudio(ByteArray(8), "composer")
-        queue.markFailure("failed-one", "Whisper auth failed")
-
-        val (vm, _) = newVmWithQueue(
-            whisper = fakeWhisperClient { Result.success("auto-retry success") },
-            samplerDispatcher = StandardTestDispatcher(testScheduler),
-            queue = queue,
-        )
-        vm.onForegroundResume()
-        advanceUntilIdle()
-
-        // Only the offline row was retried successfully.
-        assertTrue(queue.succeededIds.contains("offline-one"))
-        assertFalse(queue.succeededIds.contains("failed-one"))
-    }
-
-    @Test
-    fun foregroundResumeNoOpWhenOffline() = runTest {
-        val queue = FakePendingQueue()
-        queue.nextId = "offline-row"
-        queue.enqueueAudio(
-            ByteArray(8),
-            "composer",
-            initialError = com.pocketshell.app.voice.PendingTranscriptionItem.NETWORK_WAITING_MESSAGE,
-        )
-
-        var whisperCalls = 0
-        val whisper = object : WhisperClient {
-            override suspend fun transcribe(audio: ByteArray, language: String?): Result<String> {
-                whisperCalls++
-                return Result.success("never")
-            }
-        }
-        val (vm, _) = newVmWithQueue(
-            whisper = whisper,
-            samplerDispatcher = StandardTestDispatcher(testScheduler),
-            queue = queue,
-            connectivity = FakeConnectivity(initial = false),
-        )
-        vm.onForegroundResume()
-        advanceUntilIdle()
-        assertEquals("must not burn quota while offline", 0, whisperCalls)
     }
 
     // -- Issue #185: silence-threshold surfacing + stricter floor --------
@@ -5702,12 +4985,14 @@ class PromptComposerViewModelTest {
         // cancelled and the transcript goes into the retry queue with
         // an error message.
         val mic = FakeMicCapture()
-        val (vm, queue) = newVmWithQueue(
+        val queue = FakePendingQueue()
+        val vm = newVm(
             mic = mic,
             whisper = fakeWhisperClient {
                 Result.failure(WhisperException.Transport("simulated drop"))
             },
             samplerDispatcher = StandardTestDispatcher(testScheduler),
+            pendingTranscriptionStore = queue,
         )
         val sent = collectSendRequests(vm)
 
@@ -5742,9 +5027,11 @@ class PromptComposerViewModelTest {
                 return Result.success("never")
             }
         }
-        val (vm, queue) = newVmWithQueue(
+        val queue = FakePendingQueue()
+        val vm = newVm(
             whisper = whisper,
             samplerDispatcher = StandardTestDispatcher(testScheduler),
+            pendingTranscriptionStore = queue,
             connectivity = FakeConnectivity(initial = false),
         )
         val sent = collectSendRequests(vm)
